Show a user's name on their dashboard
When a user visits either '/admin/dashboard' or '/student/dashboard', if they are authorized to view that dashboard, they see a welcome message with their name in it. This comes from sending a GET to the server at '/api/v1/:role/dashboard', which responds with JSON containing the user's name.