Send the JSON web token in an HTTP header
This commit updates the client for a change in the backend API by sending the JSON web token in an HTTP header instead of in the query parameters. The intent is to promote single responsibility by focusing the purpose of the query parameters on resource attributes in RESTful API calls.