Implement a form for creating a teacher
An admin user who visits the admin dashboard can click on 'Create a New Teacher'. The admin user is then taken to a form where they can enter the teacher's attributes. Submitting the form sends a POST to the server and creates a user with a teacher role in the database!