Implement a form for creating a lesson
The teacher dashboard page has a link to a new lesson form. The form has title and text fields, as well as asset upload functionality. Submitting the form sends a POST to the server that responds with a pre-signed S3 asset upload URL, which the client uploads the asset to and then POSTs to the server again with the lesson title, text, and asset storage URL.