Update request for sending a new user email
When a user fills out the form for initiating the process of creating a user, they only fill out firstName, lastName, and email fields. When they submit the form, it sends a POST to the server at the path '/api/v1/send_new_user_email', which calls the server to send an email to the provided email address that the new user can use for setting their password.