Implement a set password form
When an admin starts the process of creating a new user, the new user receives an email containing a link for setting their password.

This commit implements the page under the password set link. The new user sees a form with fields for a password and password confirmation. When the user enters a matching password and password confirmation, the client POSTs to the server (which creates a User record in the database). Upon receiving a successful response from the server, the client automatically logs in using the new user's email and password, and is redirected to the appropriate dashboard for their user role.