Implement a form for initiating password reset
This commit adds a 'Forgot Password?' link to the login page. This link redirects to a form where a user can enter their email address in order to receive an email for resetting their password. When the user submits the form, the client sends a POST to the server with the email, and the user is redirected to a page with a message saying that a password reset email will be sent.