Update set password form for resets
This commit makes the password and password confirmation form work for not only the first time a user sets their password, but also for when they reset their password. If there is no 'reset' key in the query parameters, the form submits to a URL for setting a user's password for the first time; if there is a 'reset' key in the query parameters, the form submits to a URL for re-setting a user's password.