Implement a logout link
A logged in user sees a logout link at the top of the page. When a user clicks the logout link, their JSON web token with authentication data is removed from their browser storage and they see a brief page with a message 'You are being logged out!'. After a little over a second, this page redirects to the login form.