Remove teachers index page
Previously, when an admin created a new teacher, the admin would be redirected to a teachers index page. This commit simplifies the app by removing the teachers index page (because the customer has not requested it for the minimum viable product). When an admin creates a new teacher, they are now redirected to a page that says that an email has been sent to the new teacher and the new teacher will need to follow the link in the email to set their password.