Add a form for creating a student
An admin user can visit their dashboard, click the link 'Create a New Student', fill in first name / last name / email fields, and submit the form. This POSTs to the server to create a new user with a student role. If the server responds with a 204 (indicating that an initial password set email was sent to the new user's email address), the admin user is redirected to a page that says that the password set email was sent. If the server responds with any status code other than 204, the admin is redirected to a 404 page.