Enable a teacher to assign lessons to students
A teacher can visit a lesson page, and at the bottom of the page the teacher will see a list of student names with checkboxes. Teachers are the only user role who see this list when they visit a lesson page. The teacher can select any number of students and click the button 'Assign to Students'. This sends a POST to the web service endpoint for creating an assignment record that joins a lesson and a student. The teacher then sees the list of all students disappear, and sees a message saying that the assignments have been created.