Make new assignment link work predictably
When a teacher assigns a lesson to a student, the server sends an email to the student with a link to the assigned lesson. Typically, when the student clicks on that link, they would not yet be authenticated; this means that the student would be redirected from the lesson show page to the login page, and after authentication they would be redirected to their dashboard instead of to the lesson show page. This would break the user's expectation that they are visiting a lesson show page when they click on the link to a lesson.

This commit implements functionality so that if an unauthenticated user visits a lesson show page, they are redirected to the login form, and when they authenticate, they are then redirected to the lesson show page as expected.