Show assigned lessons in student dashboard
When a student visits their dashboard, they see a list of lessons assigned to th
em. Each item in this list is a link to that lesson's show page.

This commit adds a helper function 'getID' for getting the user's id out of the JSON web token, and updates the way helper functions are exported so that both the 'getID' and 'getRole' functions can be used throughout the codebase.