Tighten up display of a lesson
Previously, the content of a lesson was using the entire viewport width on a desktop device, which wasn't consistent with the rest of the app and was a bit awkward to read. The content of a lesson is now centered in the viewport on a desktop device and uses up half of the viewport. It still uses the entire viewport width on a mobile device, which seems reasonable.